Create blank.yml
Updated the GitHub Actions CI/CD workflow

This commit modifies the 'tags' field in the 'Build and push Docker image' step of our CI/CD pipeline. The change ensures that the Docker image gets pushed to the correct DockerHub repository ('first-repository') under our DockerHub username ('hmactian'). This is crucial for our deployment process as it helps automate the building and pushing of the Docker image whenever changes are pushed to the 'main' branch of our repository.